Rename `uv-traits` and split into separate modules
