+++ Cargo.toml
@@ -43,7 +43,7 @@ uv-interpreter = { path = "crates/uv-interpreter" }
 uv-normalize = { path = "crates/uv-normalize" }
 uv-requirements = { path = "crates/uv-requirements" }
 uv-resolver = { path = "crates/uv-resolver" }
-uv-traits = { path = "crates/uv-traits" }
+uv-types = { path = "crates/uv-types" }
 uv-trampoline = { path = "crates/uv-trampoline" }
 uv-version = { path = "crates/uv-version" }
 uv-virtualenv = { path = "crates/uv-virtualenv" }
