+++ crates/uv-installer/Cargo.toml
@@ -27,7 +27,7 @@ uv-extract = { workspace = true }
 uv-fs = { workspace = true }
 uv-interpreter = { workspace = true }
 uv-normalize = { workspace = true }
-uv-traits = { workspace = true }
+uv-types = { workspace = true }
 uv-warnings = { workspace = true }
 
 anyhow = { workspace = true }
