+++ crates/uv-resolver/src/resolver/provider.rs
@@ -10,7 +10,7 @@ use pypi_types::Metadata23;
 use uv_client::{FlatIndex, RegistryClient};
 use uv_distribution::DistributionDatabase;
 use uv_normalize::PackageName;
-use uv_traits::{BuildContext, NoBinary, NoBuild};
+use uv_types::{BuildContext, NoBinary, NoBuild};
 
 use crate::python_requirement::PythonRequirement;
 use crate::version_map::VersionMap;
