+++ crates/uv-resolver/src/version_map.rs
@@ -16,7 +16,7 @@ use pypi_types::{Hashes, Yanked};
 use rkyv::{de::deserializers::SharedDeserializeMap, Deserialize};
 use uv_client::{FlatDistributions, OwnedArchive, SimpleMetadata, VersionFiles};
 use uv_normalize::PackageName;
-use uv_traits::{NoBinary, NoBuild};
+use uv_types::{NoBinary, NoBuild};
 use uv_warnings::warn_user_once;
 
 use crate::{python_requirement::PythonRequirement, yanks::AllowedYanks};
