+++ crates/uv-types/src/downloads.rs
@@ -1,0 +1,8 @@
+use distribution_types::{CachedDist, DistributionId};
+use once_map::OnceMap;
+
+#[derive(Default)]
+pub struct InFlight {
+    /// The in-flight distribution downloads.
+    pub downloads: OnceMap<DistributionId, Result<CachedDist, String>>,
+}
